fshack: Generate FSR resolutions based on native aspect ratio
The base list of resolutions remains unchanged.
If FSR is enabled it adds FSR resolutions for the native resolution.
If FSR_MODE is specified, it adds only the corresponding mode.
If FSR_CUSTOM_MODE is specified, adds that to the list too.

The FSR_MODE and FSR_CUSTOM_MODE options limit resolutions to their
respective values, with FSR_CUSTOM_MODE taking precedence.